ejemplo_microservicios/microservice5/myshop/requirements.txt
django==3.1.7
djangorestframework==3.12.2
django-cors-headers==3.5.0
gunicorn==20.0.4
Pillow==8.1.1
Celery
django-environ==0.4.5